reliable_req_rep_patterns/lazy_pirate/python/requirements.txt
python-dotenv==0.19.0
pyzmq==22.3.0